refactor: change the variable name according to the situation
