+++ src/dom/update/updatePurchaseTable.js
