+++ src/dom/update/updateVendingMachineCharge.js
